Create xor-relu-22222-16.json
This is a xor network created using convnet.js as a simple rnn with 22222 nodes on 1 hidden relu layer. This network trained to 99% on all four data points in 16 iterations of all 4 data points [[0,0],[1,1],[1,0],[0,1]]  the best I was able to achieve using convnetjs in the browser. Other networks collapsed at 33333 nodes in a single hidden layer. The network can be loaded directly into a convnet for further use. I am adding this file to demonstrate the limitations of convnet and potentials of convnet in the browser (Chrome in my case.)